+++ Changes.md
@@ -22,6 +22,7 @@ API
 - PlugPopup : Added new class for editing plugs in a popup window.
 - ColorChooserPlugValueWidget : Added a new class to allow colours to be edited directly using a ColorChooser.
 - ColorChooser : Added `setSwatchesVisible()` and `getSwatchesVisible()` methods.
+- PathColumn : Added new class for adding custom columns to PathListingWidget.
 
 0.61.1.1 (relative to 0.61.1.0)
 ========